feat(catalog): We can now explicitely set permissions on copied files
This is necessary if we want to be able to copy the files directly from the resources
within the JAR files we are running from. JAR files don't retain file permissions
like you would expect from ZIP files so we have to reset the permissions ourselves.
For this a new file `fileattr.yaml` has been added.